centralize message log, will be reused for aspnet (#1020)
