+++ test/Microsoft.Azure.SignalR.Tests/Logging/MessageLogTests.cs
@@ -19,49 +19,49 @@ namespace Microsoft.Azure.SignalR.Tests
             using (var s = new ClientConnectionScope(null, true))
             {
                 // broadcast
-                ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessage(logger, new BroadcastDataMessage(null, 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageTemplate, 123UL), logger.LogStr);
+                AzureSignalRLog.StartToBroadcastMessage(logger, new BroadcastDataMessage(null, 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToBroadcastMessageTemplate, 123UL), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessage(logger, new BroadcastDataMessage(new[] { "x", "y" }, new Dictionary<string, ReadOnlyMemory<byte>>(), 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageWithExcludedConnectionTemplate, 123UL, 2, "x, y"), logger.LogStr);
+                AzureSignalRLog.StartToBroadcastMessage(logger, new BroadcastDataMessage(new[] { "x", "y" }, new Dictionary<string, ReadOnlyMemory<byte>>(), 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToBroadcastMessageWithExcludedConnectionTemplate, 123UL, 2, "x, y"), logger.LogStr);
 
                 // send to connections
-                ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToConnections(logger, new MultiConnectionDataMessage(new[] { "id1", "id2" }, null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToConnectionsTemplate, 123UL, 2, "id1, id2"), logger.LogStr);
+                AzureSignalRLog.StartToSendMessageToConnections(logger, new MultiConnectionDataMessage(new[] { "id1", "id2" }, null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToSendMessageToConnectionsTemplate, 123UL, 2, "id1, id2"), logger.LogStr);
 
                 // send to user/users
-                ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToUser(logger, new UserDataMessage("user", null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToUserTemplate, 123UL, "user"), logger.LogStr);
+                AzureSignalRLog.StartToSendMessageToUser(logger, new UserDataMessage("user", null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToSendMessageToUserTemplate, 123UL, "user"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToUsers(logger, new MultiUserDataMessage(new[] { "u1", "u2" }, null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToSendMessageToUsersTemplate, 123UL, 2, "u1, u2"), logger.LogStr);
+                AzureSignalRLog.StartToSendMessageToUsers(logger, new MultiUserDataMessage(new[] { "u1", "u2" }, null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToSendMessageToUsersTemplate, 123UL, 2, "u1, u2"), logger.LogStr);
 
                 // send to group/groups
-                ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroup(logger, new GroupBroadcastDataMessage("g", null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroupTemplate, 123UL, "g"), logger.LogStr);
+                AzureSignalRLog.StartToBroadcastMessageToGroup(logger, new GroupBroadcastDataMessage("g", null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToBroadcastMessageToGroupTemplate, 123UL, "g"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroup(logger, new GroupBroadcastDataMessage("g", new[] { "c1", "c2" }, null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroupWithExcludedConnectionsTemplate, 123UL, "g", 2, "c1, c2"), logger.LogStr);
+                AzureSignalRLog.StartToBroadcastMessageToGroup(logger, new GroupBroadcastDataMessage("g", new[] { "c1", "c2" }, null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToBroadcastMessageToGroupWithExcludedConnectionsTemplate, 123UL, "g", 2, "c1, c2"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroups(logger, new MultiGroupBroadcastDataMessage(new[] { "g1", "g2" }, null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToBroadcastMessageToGroupsTemplate, 123UL, 2, "g1, g2"), logger.LogStr);
+                AzureSignalRLog.StartToBroadcastMessageToGroups(logger, new MultiGroupBroadcastDataMessage(new[] { "g1", "g2" }, null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToBroadcastMessageToGroupsTemplate, 123UL, 2, "g1, g2"), logger.LogStr);
 
                 // connection join/leave group
-                ServiceLifetimeManagerBase<Hub>.Log.StartToAddConnectionToGroup(logger, new JoinGroupWithAckMessage("c", "g", tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToAddConnectionToGroupTemplate, 123UL, "c", "g"), logger.LogStr);
+                AzureSignalRLog.StartToAddConnectionToGroup(logger, new JoinGroupWithAckMessage("c", "g", tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToAddConnectionToGroupTemplate, 123UL, "c", "g"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveConnectionFromGroup(logger, new LeaveGroupWithAckMessage("c", "g", tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveConnectionFromGroupTemplate, 123UL, "c", "g"), logger.LogStr);
+                AzureSignalRLog.StartToRemoveConnectionFromGroup(logger, new LeaveGroupWithAckMessage("c", "g", tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToRemoveConnectionFromGroupTemplate, 123UL, "c", "g"), logger.LogStr);
 
                 // user join/leave group
-                ServiceLifetimeManagerBase<Hub>.Log.StartToAddUserToGroup(logger, new UserJoinGroupMessage("c", "g", tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToAddUserToGroupTemplate, 123UL, "c", "g"), logger.LogStr);
+                AzureSignalRLog.StartToAddUserToGroup(logger, new UserJoinGroupMessage("c", "g", tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToAddUserToGroupTemplate, 123UL, "c", "g"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveUserFromGroup(logger, new UserLeaveGroupMessage("c", "g", tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveUserFromGroupTemplate, 123UL, "c", "g"), logger.LogStr);
+                AzureSignalRLog.StartToRemoveUserFromGroup(logger, new UserLeaveGroupMessage("c", "g", tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToRemoveUserFromGroupTemplate, 123UL, "c", "g"), logger.LogStr);
 
-                ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveUserFromGroup(logger, new UserLeaveGroupMessage("c", null, tracingId: 123UL));
-                Assert.Equal(string.Format(ServiceLifetimeManagerBase<Hub>.Log.StartToRemoveUserFromAllGroupsTemplate, 123UL, "c"), logger.LogStr);
+                AzureSignalRLog.StartToRemoveUserFromGroup(logger, new UserLeaveGroupMessage("c", null, tracingId: 123UL));
+                Assert.Equal(string.Format(AzureSignalRLog.StartToRemoveUserFromAllGroupsTemplate, 123UL, "c"), logger.LogStr);
             }
         }
